+++ lib/AppStuff/app_pics.dart
@@ -1,0 +1,6 @@
+class AppPics {
+  static const String burger = 'assets/images/burger.png';
+  static const String burger1 = 'assets/images/one burger.png';
+  static const String burger2 = 'assets/images/two burger.png';
+
+}
