+++ lib/AppStuff/app_strings.dart
@@ -1,0 +1,8 @@
+class AppStrings {
+  static const String appBar = 'Krab Burger';
+  static const String declare = 'I love burgers!';
+  static const String title = 'Crusty Krab';
+  static const String copyright = 'Product made by:\nTabitha Mutinda';
+  static const String signIn = 'Sign In!';
+  static const String logIn = 'Log In!';
+}
