+++ pubspec.yaml
@@ -27,6 +27,7 @@ environment:
 # the latest version available on pub.dev. To see which dependencies have newer
 # versions available, run `flutter pub outdated`.
 dependencies:
+  local_auth: ^1.1.10
   flutter:
     sdk: flutter
 
@@ -65,11 +66,7 @@ flutter:
 
   # To add assets to your application, add an assets section, like this:
 assets:
-   - assets/
-   - assets/images/
-  #  - assets/images/burger.png
-   - assets/images/two burger.png
-
+  - assets/images/
   #   - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
